app-etl/requirements.txt
mysql-connector-python
pymssql